'Cmd+ç' also acts as select all text on NMC with a Portuguese keyboard layout https://jira.ccpgames.com/browse/EO-22434
Fixed a bad typecast

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1950063]

+++ UI/Scancodes.mm
@@ -88,7 +88,7 @@ public:
 	
 	uint32_t operator[]( char c ) const
 	{
-		return m_keyForChar[int( c )];
+		return m_keyForChar[int( (uint8_t)c )];
 	}
 private:
 	uint32_t m_keyForChar[256];
